chore(scraper): renamed module import
